Issue #42 Fix error message when width not specified
Added newline to error message.
